+++ yt/yt/core/misc/async_slru_cache-inl.h
@@ -639,7 +639,7 @@ void TAsyncSlruCacheBase<TKey, TValue, THash>::Trim(TShard* shard, NConcurrency:
     // Evict from younger.
     std::vector<TValuePtr> evictedValues;
     while (!shard->YoungerLruList.Empty() &&
-        Config_->ShardCount * shard->YoungerWeightCounter + OlderWeightCounter_.load() > capacity)
+        static_cast<i64>(Config_->ShardCount * shard->YoungerWeightCounter + OlderWeightCounter_.load()) > capacity)
     {
         auto* item = &*(--shard->YoungerLruList.End());
         auto value = item->Value;
