+++ yt/yt/ytlib/table_client/skynet_column_evaluator.cpp
@@ -176,31 +176,31 @@ void TSkynetColumnEvaluator::UnpackFields(
     TUnversionedValue** md5,
     TUnversionedValue** dataSize)
 {
-    YT_VERIFY(fullRow.GetCount() >= FilenameId_);
+    YT_VERIFY(static_cast<int>(fullRow.GetCount()) >= FilenameId_);
     if (fullRow[FilenameId_].Type != EValueType::String) {
         THROW_ERROR_EXCEPTION("Missing \"filename\" column");
     }
     *filename = TStringBuf(fullRow[FilenameId_].Data.String, fullRow[FilenameId_].Length);
 
-    YT_VERIFY(fullRow.GetCount() >= PartIndexId_);
+    YT_VERIFY(static_cast<int>(fullRow.GetCount()) >= PartIndexId_);
     if (fullRow[PartIndexId_].Type != EValueType::Int64) {
         THROW_ERROR_EXCEPTION("Missing \"part_index\" column");
     }
     *partIndex = fullRow[PartIndexId_].Data.Int64;
 
-    YT_VERIFY(fullRow.GetCount() >= DataId_);
+    YT_VERIFY(static_cast<int>(fullRow.GetCount()) >= DataId_);
     if (fullRow[DataId_].Type != EValueType::String) {
         THROW_ERROR_EXCEPTION("Missing \"data\" column");
     }
     *data = TStringBuf(fullRow[DataId_].Data.String, fullRow[DataId_].Length);
 
-    YT_VERIFY(fullRow.GetCount() >= Sha1Id_);
+    YT_VERIFY(static_cast<int>(fullRow.GetCount()) >= Sha1Id_);
     *sha1 = &fullRow[Sha1Id_];
 
-    YT_VERIFY(fullRow.GetCount() >= Md5Id_);
+    YT_VERIFY(static_cast<int>(fullRow.GetCount()) >= Md5Id_);
     *md5 = &fullRow[Md5Id_];
 
-    YT_VERIFY(fullRow.GetCount() >= DataSizeId_);
+    YT_VERIFY(static_cast<int>(fullRow.GetCount()) >= DataSizeId_);
     *dataSize = &fullRow[DataSizeId_];
 }
